Dynamic light functionality
Day/Night light cycle, proof of concept. This is by no means a complete project but it's a good starting point to play around with! Check out the TimeTracker script attached to the _TimeTracker gameobject in the hierarchy. Are there ways that you can think of to extend and expand this funtionality?